+++ core-ng/src/test/java/core/framework/impl/web/bean/QueryParamMapperTest.java
@@ -1,26 +1,34 @@
 package core.framework.impl.web.bean;
 
+import core.framework.util.ClasspathResources;
 import core.framework.util.Maps;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Map;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author neo
  */
 class QueryParamMapperTest {
+    private QueryParamMapperBuilder<TestQueryParamBean> builder;
     private QueryParamMapper<TestQueryParamBean> mapper;
 
     @BeforeEach
     void createMapper() {
-        mapper = new QueryParamMapperBuilder<>(TestQueryParamBean.class).build();
+        builder = new QueryParamMapperBuilder<>(TestQueryParamBean.class);
+        mapper = builder.build();
+    }
+
+    @Test
+    void sourceCode() {
+        String sourceCode = builder.builder.sourceCode();
+        assertEquals(ClasspathResources.text("query-param-mapper-test/test-bean-mapper.java"), sourceCode);
     }
 
     @Test
@@ -34,11 +42,12 @@ class QueryParamMapperTest {
 
         Map<String, String> params = mapper.toParams(bean);
 
-        assertEquals("12", params.get("int_field"));
-        assertEquals("22.3", params.get("double_field"));
-        assertEquals("value", params.get("string_field"));
-        assertEquals("2017-08-28T13:44:00", params.get("date_time_field"));
-        assertEquals("V2", params.get("enum_field"));
+        assertThat(params).containsEntry("int_field", "12")
+                          .containsEntry("double_field", "22.3")
+                          .containsEntry("string_field", "value")
+                          .containsEntry("date_time_field", "2017-08-28T13:44:00")
+                          .containsEntry("enum_field", "V2")
+                          .containsEntry("default_value_field", "value");
     }
 
     @Test
@@ -51,9 +60,10 @@ class QueryParamMapperTest {
 
         TestQueryParamBean bean = mapper.fromParams(params);
 
-        assertTrue(bean.booleanField);
-        assertEquals(BigDecimal.valueOf(345.67), bean.bigDecimalField);
-        assertEquals(LocalDate.of(2017, 8, 28), bean.dateField);
-        assertEquals(Long.valueOf(123), bean.longField);
+        assertThat(bean.booleanField).isTrue();
+        assertThat(bean.bigDecimalField).isEqualTo("345.67");
+        assertThat(bean.dateField).isEqualTo(LocalDate.of(2017, 8, 28));
+        assertThat(bean.longField).isEqualTo(123);
+        assertThat(bean.defaultValueField).isEqualTo("value");
     }
 }